+++ um-webapp/src/main/java/gov/nara/um/persistence/dao/IBusinessUnitDao.java
@@ -5,7 +5,7 @@ import gov.nara.um.persistence.model.BusinessUnit;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface IBusinessUnitDao extends JpaRepository<BusinessUnit, Long>, JpaSpecificationExecutor<BusinessUnit>, IByNameApi<BusinessUnit> {
+public interface IBusinessUnitDao extends JpaRepository<BusinessUnit, Integer>, JpaSpecificationExecutor<BusinessUnit>, IByNameApi<BusinessUnit> {
 
 
 }
